requirements.txt
Pillow
Colorama
Streamlit
